refactor(im): lift CardState to shared im/card_model.py
Move CardState, ToolStep, ArtifactItem, PendingInput, SubAgentRow from
im/feishu/card_model.py into im/card_model.py so Discord (and future
platforms) can use card state without importing from feishu. The feishu
module re-exports all five classes for backward compatibility; existing
importers in outbound.py, artifacts.py, and types.py are updated to use
the canonical location.
